Separate out CameraVis from CameraVisCoordinator
